Embed into Python
Adds Embedding LlamaScript into Python, making LlamaScript more powerful.
